reorganized selectors
