+++ projects/ngx-object-diagram/src/lib/state/graph.reducer.ts
@@ -1,4 +1,4 @@
-import { createFeatureSelector, createReducer, createSelector, on } from "@ngrx/store";
+import { createReducer, on } from "@ngrx/store";
 import * as GraphActions from "./graph.actions"
 import { Entity } from "../model/entity";
 import { immerOn } from "ngrx-immer/store";
@@ -37,31 +37,3 @@ export const graphReducer = createReducer<GraphState>(
     state.graphs[state.currentGraphId]!.find(e => e.guid == entity.guid)!.collapsed = !entity.collapsed;
   })
 )
-
-export const selectGraphFeatureState = createFeatureSelector<GraphState>((() => {
-  console.log("k")
-  return 'graph';
-})());
-
-export const selectCurrentGraphId = createSelector(
-  selectGraphFeatureState,
-  state => state.currentGraphId
-);
-
-export const selectEntities = createSelector(
-  selectGraphFeatureState,
-  selectCurrentGraphId,
-  (state, graphId) => {
-    return state?.graphs?.[graphId];
-  }
-);
-
-export const selectEntity = (entityGuid: string) =>
-  createSelector(selectEntities,
-    entities => entities?.find(e => e.guid == entityGuid)
-);
-
-export const selectCollapsed = (entityGuid: string) =>
-  createSelector(selectEntity(entityGuid),
-    entity => entity?.collapsed
-);
